Fix SSLInterceptor NPE
The spec states returning a non-null (possibly empty) array of issuers
and whilst a null works on the default it causes a NPE in some
implementations.
Whilst I can’t find doc that clearly states it behaviour suggests that
the empty array is a special case that doesn’t mean trust none, but
trust all.

Parameterised SSL tests for both HTTP types to ensure that we get
SSL coverage of both implementations and don't regress one or the
other.
Moved OkFactoryBlocker and associated parameterisation functions to
test utils.
Made SslAuthenticationTest also use both okhttp and the default
HttpURLConnection.